+++ packages/snaps-execution-environments/src/common/validation.ts
@@ -1,5 +1,5 @@
 import { ChainIdStruct, HandlerType } from '@metamask/snaps-utils';
-import { SnapKeyring } from '@metamask/snap-types';
+import { SnapKeyring } from '@metamask/snaps-types';
 import {
   array,
   assign,