+++ packages/snaps-types/CHANGELOG.md
@@ -38,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BREAKING:** Add Transaction Insight API ([#642](https://github.com/MetaMask/snaps-monorepo/pull/642))
 
 ### Removed
-- **BREAKING:** Move all internal types from `@metamask/snap-types` to `@metamask/snaps-utils` ([#695](https://github.com/MetaMask/snaps-monorepo/pull/695))
+- **BREAKING:** Move all internal types from `@metamask/snaps-types` to `@metamask/snaps-utils` ([#695](https://github.com/MetaMask/snaps-monorepo/pull/695))
   - Previously accessible types can now be accessed by importing `@metamask/snaps-utils`
 
 ## [0.19.1]
@@ -190,7 +190,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Readme file ([#71](https://github.com/MetaMask/snaps-monorepo/pull/71))
 
 ### Changed
-- **BREAKING:** Rename package to `@metamask/snap-types` ([#73](https://github.com/MetaMask/snaps-monorepo/pull/73))
+- **BREAKING:** Rename package to `@metamask/snaps-types` ([#73](https://github.com/MetaMask/snaps-monorepo/pull/73))
 
 ## [0.0.6]